refactor: share OLAP engine label logic between web-local and web-admin
Move `getOlapEngineLabel` to `web-common/features/resources/display-utils.ts`
and re-export from web-admin. Relax the helper so callers that only have a
connector name (like web-local's `instance.olapConnector`) still get DuckLake
and basic type formatting. Web-local's status page now uses the same helper,
so a DuckLake connector displays as "DuckLake" in both Rill Developer and
Rill Cloud.
