+++ web-common/src/features/resources/display-utils.ts
@@ -1,3 +1,5 @@
+import type { V1Connector } from "@rilldata/web-common/runtime-client";
+
 /**
  * Formats a connector name for display with proper capitalization.
  * Handles known connectors (duckdb, clickhouse, etc.) with correct casing.
@@ -17,6 +19,51 @@ export function formatConnectorName(connector: string | undefined): string {
   if (lower === "claude") return "Claude";
   if (lower === "gemini") return "Gemini";
   return connector.charAt(0).toUpperCase() + connector.slice(1);
+}
+
+/**
+ * Returns the display label for the OLAP engine, including MotherDuck and
+ * DuckLake detection and a management suffix (Rill-managed) where applicable.
+ *
+ * MotherDuck is detected by the connector's path starting with "md:" or a
+ * token being configured. DuckLake is detected by `config.attach` referencing
+ * a DuckLake catalog, or (as a fallback when configs are redacted or callers
+ * only have the connector name) by a `ducklake` / `ducklake_*` name.
+ *
+ * Accepts a full V1Connector (with type + config) or a partial object holding
+ * just a name — the latter is what web-local has access to via
+ * `instance.olapConnector`.
+ */
+export function getOlapEngineLabel(connector: V1Connector | undefined): string {
+  if (!connector) return "DuckDB";
+
+  const name = connector.name ?? "";
+  const lowerName = name.toLowerCase();
+  const type = connector.type ?? "";
+
+  const isDuckDB = type === "duckdb";
+  const isMotherDuck =
+    isDuckDB &&
+    (String(connector.config?.path ?? "").startsWith("md:") ||
+      !!connector.config?.token);
+  const isDuckLake =
+    !isMotherDuck &&
+    ((isDuckDB &&
+      String(connector.config?.attach ?? "").includes("ducklake:")) ||
+      lowerName === "ducklake" ||
+      lowerName.startsWith("ducklake_"));
+
+  // When `type` is missing (e.g. web-local only has the olap connector name),
+  // fall back to the name so "duckdb"/"clickhouse" still format correctly.
+  const resolvedType = isMotherDuck
+    ? "motherduck"
+    : isDuckLake
+      ? "ducklake"
+      : type || name;
+  const label = formatConnectorName(resolvedType);
+
+  if (connector.provision) return `${label} (Rill-managed)`;
+  return label;
 }
 
 /**
